[SYCLomatic][NFC] Move pre-processor code into seperate PreProcessor.cpp/h files.
Remove unnecessary includes.
