+++ clang/lib/DPCT/MigrationAction.h
@@ -12,7 +12,9 @@
 #include "clang/Frontend/FrontendAction.h"
 #include "clang/Tooling/Tooling.h"
 
+#include "PreProcessor.h"
 #include "ASTTraversal.h"
+
 #include "RulesSecurity/MisleadingBidirectional.h"
 
 namespace clang {
