+++ clang/lib/DPCT/RulesInclude/InclusionHeaders.cpp
@@ -7,7 +7,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "InclusionHeaders.h"
-
+#include "PreProcessor.h"
 #include "ASTTraversal.h"
 
 namespace clang {
